requirements.txt
asgiref==3.3.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
idna==2.10
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
pyuploadcare==2.7.0
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
